refactor: dissolve pcg32.ts into createPcg.ts and types.ts
The number-based PCG32 implementation lived in its own file. Move:
- PCGState32 and RandomFn types -> types.ts (alongside PCGState64/RandomFn64)
- createPcg32, nextState, prevState, stepState, getOutput, randomInt,
  randomUint32, randomList (plus the SETSEQ/XSH_RR constants and
  setseqIncrement helper) -> createPcg.ts (above the BigInt PCG64 factory)

The source tree is now createPcg / types / uint64 (plus bitwise/defaults/
index). pcg32.ts is deleted. Imports in index.ts, the bench, and the
test are repointed.

No algorithm changes; bench checksums match (2195024581, 312469475).
